Sync datatype definitions (#11)

+++ src/Zenon/Model/Embedded/Json/JPillarInfo.cs
@@ -3,19 +3,19 @@
     public class JPillarInfo
     {
         public string name { get; set; }
-        public long rank { get; set; }
+        public int rank { get; set; }
         public int type { get; set; }
         public string ownerAddress { get; set; }
         public string producerAddress { get; set; }
         public string withdrawAddress { get; set; }
-        public long giveMomentumRewardPercentage { get; set; }
-        public long giveDelegateRewardPercentage { get; set; }
+        public int giveMomentumRewardPercentage { get; set; }
+        public int giveDelegateRewardPercentage { get; set; }
         public bool isRevocable { get; set; }
         public long revokeCooldown { get; set; }
         public long revokeTimestamp { get; set; }
         public JPillarEpochStats currentStats { get; set; }
         public string weight { get; set; }
-        public long producedMomentums { get; set; }
-        public long expectedMomentums { get; set; }
+        public int producedMomentums { get; set; }
+        public int expectedMomentums { get; set; }
     }
 }
